Remove `Default` implementations and add a `DefaultToHost` utility.
The `Default` implementations here were inspired by target triple
parsing, where the various fields default to "unknown", however they
doesn't make as much sense outside of that context. So, remove the
`Default` implementations, and add a separate `DefaultToHost` for
use by users that want a `Triple` that defaults to targeting the host.

Fixes #49.